% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-assay.R
\name{CreateSCleanAssayFromHDF5}
\alias{CreateSCleanAssayFromHDF5}
\title{Create a scLeanAssay from an existing scLean HDF5 file}
\usage{
CreateSCleanAssayFromHDF5(hdf5_path, assay = "RNA")
}
\arguments{
\item{hdf5_path}{Path to the HDF5 file}

\item{assay}{Assay name (default: "RNA")}
}
\value{
A scLeanAssay object
}
\description{
Re-opens a previously created HDF5-backed assay. All layers found in the
HDF5 file under \code{/assays/<assay>/layers/} are wrapped as
HDF5BackedMatrix proxies.
}
\examples{
\donttest{
counts <- Matrix::rsparsematrix(100, 50, density = 0.1,
  dimnames = list(paste0("gene", 1:100), paste0("cell", 1:50)))
tmp <- tempfile(fileext = ".h5")
CreateSCleanAssay(counts, hdf5_path = tmp)
assay <- CreateSCleanAssayFromHDF5(tmp)
assay
}
}
\seealso{
\code{\link{CreateSCleanAssay}}, \code{\link{LoadScleanObject}}
}
