% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5-matrix.R
\name{HDF5BackedMatrix}
\alias{HDF5BackedMatrix}
\alias{nrow.HDF5BackedMatrix}
\alias{ncol.HDF5BackedMatrix}
\title{HDF5BackedMatrix: A matrix proxy backed by HDF5 storage}
\usage{
nrow.HDF5BackedMatrix(x)

ncol.HDF5BackedMatrix(x)
}
\arguments{
\item{x}{An HDF5BackedMatrix object}
}
\description{
This R6 class implements the standard R matrix interface (dim, dimnames, [, [<-, t, %*%)
but reads/writes data from HDF5 on the fly. Designed to be a drop-in replacement
for dgCMatrix in Seurat layers.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{hdf5_path}}{Path to the HDF5 file}

    \item{\code{group_path}}{HDF5 group containing the CSC matrix datasets}

    \item{\code{dims}}{Dimensions of the matrix}

    \item{\code{dimnames}}{Row and column names

Create a new HDF5BackedMatrix}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-HDF5BackedMatrix-initialize}{\code{HDF5BackedMatrix$new()}}
    \item \href{#method-HDF5BackedMatrix-dim}{\code{HDF5BackedMatrix$dim()}}
    \item \href{#method-HDF5BackedMatrix-nrow}{\code{HDF5BackedMatrix$nrow()}}
    \item \href{#method-HDF5BackedMatrix-ncol}{\code{HDF5BackedMatrix$ncol()}}
    \item \href{#method-HDF5BackedMatrix-get}{\code{HDF5BackedMatrix$get()}}
    \item \href{#method-HDF5BackedMatrix-as_matrix}{\code{HDF5BackedMatrix$as_matrix()}}
    \item \href{#method-HDF5BackedMatrix-t}{\code{HDF5BackedMatrix$t()}}
    \item \href{#method-HDF5BackedMatrix-mat_mult}{\code{HDF5BackedMatrix$mat_mult()}}
    \item \href{#method-HDF5BackedMatrix-clone}{\code{HDF5BackedMatrix$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-initialize}{}}}
\subsection{\code{HDF5BackedMatrix$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$new(hdf5_path, group_path)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{hdf5_path}}{Path to HDF5 file}
      \item{\code{group_path}}{Group path within the file

Get dimensions}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-dim"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-dim}{}}}
\subsection{\code{HDF5BackedMatrix$dim()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$dim()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer vector of length 2

Get number of rows
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-nrow"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-nrow}{}}}
\subsection{\code{HDF5BackedMatrix$nrow()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$nrow()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer

Get number of columns
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-ncol"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-ncol}{}}}
\subsection{\code{HDF5BackedMatrix$ncol()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$ncol()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer

Subset rows and columns
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-get"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-get}{}}}
\subsection{\code{HDF5BackedMatrix$get()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$get(i, j, drop = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{i}}{Row indices}
      \item{\code{j}}{Column indices}
      \item{\code{drop}}{Whether to drop dimensions}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A dense matrix or vector

Convert to in-memory matrix
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-as_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-as_matrix}{}}}
\subsection{\code{HDF5BackedMatrix$as_matrix()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$as_matrix()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A dgCMatrix

Transpose (lazy: returns a new proxy with swapped dims)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-t"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-t}{}}}
\subsection{\code{HDF5BackedMatrix$t()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$t()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    HDF5BackedMatrix with transposed access

Matrix multiplication
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-mat_mult"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-mat_mult}{}}}
\subsection{\code{HDF5BackedMatrix$mat_mult()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$mat_mult(x)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{x}}{Another matrix or vector}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The product
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HDF5BackedMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HDF5BackedMatrix-clone}{}}}
\subsection{\code{HDF5BackedMatrix$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HDF5BackedMatrix$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
