% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{NormalizeData}
\alias{ScaleData}
\alias{FindVariableFeatures}
\alias{RunPCA}
\alias{FindNeighbors}
\alias{FindClusters}
\alias{FindMarkers}
\alias{FindAllMarkers}
\alias{VariableFeatures}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Seurat}{\code{\link[Seurat:FindAllMarkers]{FindAllMarkers()}}, \code{\link[Seurat:FindClusters]{FindClusters()}}, \code{\link[Seurat:FindMarkers]{FindMarkers()}}, \code{\link[Seurat:FindNeighbors]{FindNeighbors()}}, \code{\link[Seurat:FindVariableFeatures]{FindVariableFeatures()}}, \code{\link[Seurat:NormalizeData]{NormalizeData()}}, \code{\link[Seurat:RunPCA]{RunPCA()}}, \code{\link[Seurat:ScaleData]{ScaleData()}}}

  \item{SeuratObject}{\code{\link[SeuratObject:VariableFeatures]{VariableFeatures()}}}
}}

