% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclean-class.R
\docType{class}
\name{scLeanAssay-class}
\alias{scLeanAssay-class}
\title{scLeanAssay: A disk-backed single-cell assay}
\description{
Extends SeuratObject::Assay5. All layer data is stored in HDF5 files
rather than in-memory R matrices.
}
\section{Slots}{

\describe{
\item{\code{hdf5_path}}{Path to the backing HDF5 file}

\item{\code{hdf5_group}}{HDF5 group path within the file for this assay}
}}

