% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-control.R
\name{SetVerbose}
\alias{SetVerbose}
\title{Toggle verbose output}
\usage{
SetVerbose(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to print progress messages}
}
\value{
The verbose setting, invisibly
}
\description{
Toggle verbose output
}
\examples{
SetVerbose(TRUE)
SetVerbose(FALSE)
}
\seealso{
\code{\link{MemoryReport}}
}
