% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclean-class.R
\name{show,scLeanAssay-method}
\alias{show,scLeanAssay-method}
\title{Show method for scLeanAssay}
\usage{
\S4method{show}{scLeanAssay}(object)
}
\arguments{
\item{object}{A scLeanAssay object}
}
\description{
Show method for scLeanAssay
}
