% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-control.R
\name{SetChunkSize}
\alias{SetChunkSize}
\title{Set chunk size for the next operation}
\usage{
SetChunkSize(object, size)
}
\arguments{
\item{object}{A Seurat/scLean object}

\item{size}{Number of rows/columns per chunk (NA to restore auto)}
}
\value{
The object, invisibly
}
\description{
Set chunk size for the next operation
}
\examples{
counts <- Matrix::rsparsematrix(100, 50, density = 0.1)
obj <- LoadScleanObject(counts)
SetChunkSize(obj, 1000)
SetChunkSize(obj, NA)  # restore auto
}
\seealso{
\code{\link{SetMaxRAM}}, \code{\link{SetThreads}}
}
