% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-control.R
\name{SetMaxRAM}
\alias{SetMaxRAM}
\title{Set maximum RAM scLean is allowed to use (in MB)}
\usage{
SetMaxRAM(max_ram = 2048)
}
\arguments{
\item{max_ram}{Memory limit in megabytes (default: 2048)}
}
\value{
The memory limit in bytes, invisibly
}
\description{
Set maximum RAM scLean is allowed to use (in MB)
}
\examples{
SetMaxRAM(4096)
SetMaxRAM()  # restore default 2048 MB
}
\seealso{
\code{\link{SetChunkSize}}, \code{\link{MemoryReport}}
}
